requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.9
scikit-image>=0.21
scikit-learn>=1.3
matplotlib>=3.7
numpy>=1.24
pillow>=10.0
shap>=0.45
